# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores/{storeId}/inventory": {
      "get": {
        "operationId": "InventoryController_listInventory",
        "summary": "List all inventory",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InventoryController_setInventory",
        "summary": "Set initial inventory",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/inventory/warehouse/{warehouseId}": {
      "get": {
        "operationId": "InventoryController_listByWarehouse",
        "summary": "List inventory for a warehouse",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/inventory/variant/{variantId}": {
      "get": {
        "operationId": "InventoryController_listByVariant",
        "summary": "List inventory for a variant",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/inventory/nearest-warehouse": {
      "get": {
        "operationId": "InventoryController_findNearestWarehouse",
        "summary": "Find nearest warehouse with stock",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/inventory/{id}": {
      "patch": {
        "operationId": "InventoryController_adjustInventory",
        "summary": "Adjust inventory quantity",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/inventory/transfer": {
      "post": {
        "operationId": "InventoryController_transferInventory",
        "summary": "Transfer stock between warehouses",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/stores/{storeId}/promo-codes": {
      "post": {
        "operationId": "DiscountsController_createPromoCode",
        "summary": "Create a promo code",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Promo code created"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DiscountsController_listPromoCodes",
        "summary": "List promo codes",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of promo codes"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/promo-codes/{id}": {
      "get": {
        "operationId": "DiscountsController_getPromoCode",
        "summary": "Get promo code details",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code details"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "DiscountsController_updatePromoCode",
        "summary": "Update promo code",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code updated"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DiscountsController_deletePromoCode",
        "summary": "Soft delete promo code",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code deactivated"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/promo-codes/{code}/validate": {
      "post": {
        "operationId": "DiscountsController_validatePromoCode",
        "summary": "Validate a promo code against a cart",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result"
          }
        },
        "tags": [
          "Promo Codes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ShopBuilder API",
    "description": "ShopBuilder — Shopify Clone for Local Kazakh Merchants. E-commerce platform API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste JWT access token (without \"Bearer \" prefix)"
      }
    },
...
  }
}